+++ parrot/src/pfs_service_local.cc
@@ -20,6 +20,7 @@ extern "C" {
 #include <string.h>
 #include <stdio.h>
 #include <errno.h>
+#include <inttypes.h>
 #include <sys/stat.h>
 #include <sys/types.h>
 #include <sys/mman.h>
@@ -670,7 +671,7 @@ public:
 		debug(D_LOCAL,"locate %s",name->rest);
 		result = stat(name, &buf);
 		if(result < 0) return 0;
-		snprintf(path, PFS_PATH_MAX, "localhost:dev%lld:%s", buf.st_dev, name->path);
+		snprintf(path, PFS_PATH_MAX, "localhost:dev%lld:%s",  (long long int) buf.st_dev, name->path);
 		loc = new pfs_location();
 		loc->append(path);
 		return loc;
